backend/pyproject.toml
[tool.poetry]
name = "task-for-rus-hydro"
version = "0.1.0"
description = "Тестовое задание для РусГидро"
authors = ["Илья Смелов <https://github.com/PivnoyFei>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.2"
pandas = "^1.5.3"
xlsxwriter = "^3.0.9"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
skip-string-normalization = true
line-length = 88
